Remove dead code
The OGLViewer was originally meant to display content
rendered with OpenGL in the WxDemo. It broke at some point
shortly after being written, and it's not used anymore.

The class also depended on an still-unreleased version of WxWidgets
which made compiling difficult, as WxWidgets had to be built from source
with FP3D. This change makes it possible to use the standard packages
for WxWidgets 3 when compiling the app.